+++ api/src/main/java/marquez/MarquezApp.java
@@ -37,9 +37,9 @@ import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
 import marquez.cli.SeedCommand;
 import marquez.db.DbMigration;
-import marquez.sentry.TracingContainerResponseFilter;
-import marquez.sentry.TracingSQLLogger;
-import marquez.sentry.TracingServletFilter;
+import marquez.tracing.TracingContainerResponseFilter;
+import marquez.tracing.TracingSQLLogger;
+import marquez.tracing.TracingServletFilter;
 import org.flywaydb.core.api.FlywayException;
 import org.jdbi.v3.core.Jdbi;
 import org.jdbi.v3.core.statement.SqlLogger;
